Add entity eid recycling
Pretty sure there is a leak somewhere at the moment.  If I start with 2k ships, I should end with 2k ships but usually end up with 1999 or 1996 or something similar.  Need to see if we can create a stress test that just creates and kills a bunch of entities in several background workers to verify the logic actually works.